разбиение кода в файле main.js на модули
